[miniflare] Prototype test file restructuring for emails
